Convert the code into object oriented
The whole source was written in procedural method. Change the code
to use classes. This is not an atomic change.

The Card struct became a Card class, and all the functions that used
it now became member functions. Similarly, the Card array deck now
became the member of a Deck class, and the related functions are wrapped
up in it. The only additions are the constructors of both classes.

The pointer *cardPtr that pointed to the top most Card in the deck array
has been scrapped. The member m_cardIndex of the Deck class holds this
information now. The method dealCard() takes care of dealing the cards
and incrementing m_cardIndex.

Signed-off: Shiva Prasad shiva.prasadv@yahoo.com